gfs2_edit: Fix savemeta test failures in 32-bit environments
size_t is the wrong type for sm_fs_bytes and it is too small in 32-bit
environments leading to savemeta test failures, as it considers the
target to be too small. Change it to uint64_t.

+++ gfs2/edit/savemeta.c
@@ -43,7 +43,7 @@ struct savemeta_header {
 struct savemeta {
 	time_t sm_time;
 	unsigned sm_format;
-	size_t sm_fs_bytes;
+	uint64_t sm_fs_bytes;
 };
 
 struct saved_metablock {
